+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+0.X.X / 2019-XX-XX
+==================
+  * api converted to use async/await with 1.39.0
+
 0.17.0 / 2019-10-22
 ==================
   * Add support for oidc providerss with `auth-provider` w/o `access-token` - #70